requirements.txt
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# Model zoo — pretrained weights from HuggingFace Hub
timm>=0.9.0

# Data & ML utilities
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.2.0
Pillow>=9.0.0

# Experiment tracking
wandb>=0.15.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bar
tqdm>=4.65.0
albumentations